+++ pkg/util/chunking_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestForEachChunk(t *testing.T) {
 	for _, datasize := range []int{0, 1, 5, 10, 50, 100, 250} {
-		for _, chunksize := range []uint64{1, 2, 3, 5, 10, 50} {
+		for _, chunksize := range []uint16{1, 2, 3, 5, 10, 50} {
 			t.Run(fmt.Sprintf("test-%d-%d", datasize, chunksize), func(t *testing.T) {
 				data := []int{}
 				for i := 0; i < datasize; i++ {
